feat: Add compile-html task module
